+++ mpicbg/src/main/java/mpicbg/ij/stack/InverseTransformMapping.java
@@ -25,7 +25,7 @@ public class InverseTransformMapping< T extends InverseCoordinateTransform > ext
 	@Override
 	public void map( final ImageStack source, final ImageProcessor target )
 	{
-		final float[] t = new float[ 3 ];
+		final double[] t = new double[ 3 ];
 		final int sw = source.getWidth() - 1;
 		final int sh = source.getHeight() - 1;
 		final int sd = source.getSize();
@@ -63,7 +63,7 @@ public class InverseTransformMapping< T extends InverseCoordinateTransform > ext
 	@Override
 	public void mapInterpolated( final ImageStack source, final ImageProcessor target )
 	{
-		final float[] t = new float[ 3 ];
+		final double[] t = new double[ 3 ];
 		final int sw = source.getWidth() - 1;
 		final int sh = source.getHeight() - 1;
 		final int sd = source.getSize();
