+++ mpicbg/src/main/java/mpicbg/models/InverseBoundable.java
@@ -1,20 +1,35 @@
+/**
+ * License: GPL
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU General Public License 2
+ * as published by the Free Software Foundation.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA  02111-1307, USA.
+ */
 package mpicbg.models;
 
 /**
  * An {@link InverseCoordinateTransform} that, for a given source interval in
  * <i>n</i>-space, can estimate the target interval in <i>n</i>-space.
- * 
- * @author Stephan Saalfeld <saalfeld@mpi-cbg.de>
- * @version 0.1b
+ *
+ * @author Stephan Saalfeld <saalfelds@janelia.hhmi.org>
  */
 public interface InverseBoundable extends InverseCoordinateTransform
 {
 	/**
 	 * Estimate the bounds of an n-dimensional interval [min,max] with min and
 	 * max being n-dimensional vectors.
-	 * 
+	 *
 	 * @param min
 	 * @param max
 	 */
-	public void estimateInverseBounds( final float[] min, final float[] max ) throws NoninvertibleModelException;
+	public void estimateInverseBounds( final double[] min, final double[] max ) throws NoninvertibleModelException;
 }
